% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragr-package.R
\docType{package}
\name{fragr-package}
\alias{fragr}
\alias{fragr-package}
\title{fragr: an R package for wrangling capillary electrophoresis data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

an R package for wrangling capillary electrophoresis data
}
\author{
\strong{Maintainer}: Eric Hunt \email{52219938+eric-hunt@users.noreply.github.com} (0000-0002-6891-8294)

}
\keyword{internal}
