% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge_prodify}
\alias{munge_prodify}
\title{Put relative measurements in terms of product formed}
\usage{
munge_prodify(nested_df, substrate_cutoff)
}
\arguments{
\item{nested_df}{a nested data frame of CE data with all peak data lying in the list column 'data'}

\item{substrate_cutoff}{a numeric value indicating the bp size above which all peaks should represent product}
}
\value{
returns the data frame relative data expressed in terms of percent product formed,
(i.e. 100% substrate will now be 0% product)
}
\description{
\code{munge_prodify}
}
