% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge_subfix}
\alias{munge_subfix}
\title{Fix instances where substrate is not defined by a single peak}
\usage{
munge_subfix(nested_df)
}
\arguments{
\item{nested_df}{a nested data frame of CE data with all peak data lying in the list column 'data'}
}
\value{
returns the data frame with the product peak(s) merged into one peak
}
\description{
\code{munge_subfix}
}
