% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_activity}
\alias{parse_activity}
\title{Determine activity according to product formed}
\usage{
parse_activity(
  nested_df,
  reg_vars,
  reg_limit,
  substrate_cutoff,
  top_n = 5,
  window_size = 2
)
}
\arguments{
\item{nested_df}{a nested data frame of CE data with all peak data lying in the list column 'data'}

\item{reg_vars}{a named vector of numeric values from \code{lm} capillary migration model,
"slope" and "intercept" must be present}

\item{reg_limit}{a numeric value defining how far the product peak may deviate from the model}

\item{substrate_cutoff}{a numeric value indicating the bp size above which all peaks should represent product}

\item{top_n}{a numeric value for noise reduction selecting the top number of peaks to analyze by peak area,
defaults to five (5) unless otherwise specified}

\item{window_size}{a numeric value indicating the ±window around the specific peak where off-target can be called on the same strand,
defaults to two (2) bp above and below "act_bp" unless otherwise specified}
}
\value{
returns the nested data frame with added variables "act", "act_bp", "offsense",
and "offanti"
}
\description{
\code{parse_activity} will determine specific and non-specific activity, expressed in terms of product formed,
uisng a capillary migration model calculated by \code{fragr::model_migration}
}
\details{
The regression model variables are used to create a predictive function,
which allows \code{parse_activity} to select the peak closest to the theoretical or predicted size
and label this peak as the specific or expected activity in relation to all activity.
The "act" variable represents the activity determined from a peak closest to the theoretical predicted size,
which is determined by the regression model variables and alllowed deviation reg_limit arguments.
The "act_bp" variable identifies the called size of the peak which was used to generate the "act" variable.
The "offsense" variable represents the sum of all other activity besides "act" on the same strand.
The "offanti" variable is only included in situations where the substate is double-stranded,
and represents the sum of all activity falling on the opposite strand as "act" and "offsense".

\code{parse_activity}
}
\seealso{
[fragr::model_migration]
}
