% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{parse_migmodel}
\alias{parse_migmodel}
\title{Generate regression model for CE migration}
\usage{
parse_migmodel(df_list, channels, search_var, search_pattern,
  substrate_cutoff, limit = 10)
}
\arguments{
\item{df_list}{a list of data frames containing CE data imported with \code{fragr::read_PeakScanner}}

\item{channels}{the one letter channel/dye code (as a character vector) for which to build a migration regression model}

\item{search_var}{a metadata-containing variable which contains the theoretical product size}

\item{search_pattern}{a regex pattern for extracting the theoretical product size 'search_var'}

\item{substrate_cutoff}{a numeric value indicating the bp size above which all peaks should represent product}

\item{limit}{a numeric value defining how far, in basepairs, the called value an deviate from the theoretical value,
defaults to ten (10) unless otherwise specified}
}
\value{
returns a model plot and list of variables..
}
\description{
Generate regression model for CE migration
}
\details{
\code{parse_migmodel}
}
