% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_PeakScanner}
\alias{read_PeakScanner}
\title{Import PeakScanner *combined table* .csv files into R}
\usage{
read_PeakScanner(directory_path, pattern = NULL, clean = TRUE)
}
\arguments{
\item{directory_path}{a path to a directory containing the .csv files}

\item{pattern}{a regex pattern for selecing files in the directory,
defaults to reading all .csv files present}

\item{clean}{an opinionated removal of some default PeakScanner variables,
defaults to TRUE}
}
\value{
generates a named list of tibbles,
each neamed element is one *combined table* named with its origin file path
}
\description{
\code{read_PeakScanner}
}
