% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_PeakScanner_file}
\alias{read_PeakScanner_file}
\title{Import single PeakScanner *combined table* .csv file into R}
\usage{
read_PeakScanner_file(file_path, clean = TRUE)
}
\arguments{
\item{file_path}{a path to a directory containing the .csv files}

\item{clean}{an opinionated removal of some default PeakScanner variables,
defaults to TRUE}
}
\value{
generates a named list of tibbles,
each neamed element is one *combined table* named with its origin file path
}
\description{
\code{read_PeakScanner_file}
}
