% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_PeakScanner_file}
\alias{read_PeakScanner_file}
\title{Import single PeakScanner *combined table* .csv file into R}
\usage{
read_PeakScanner_file(file_path, type = "legacy")
}
\arguments{
\item{file_path}{a path to a directory containing the .csv files}

\item{type}{a character string, either 'legacy' or 'cloud' indicating where
the file was generated (desktop app or online app); defaults to 'legacy'}
}
\value{
generates a named list of tibbles,
each neamed element is one *combined table* named with its origin file path
}
\description{
\code{read_PeakScanner_file}
}
